terminal: print something  when continue errors even if we don't have a file
We should print something when we exit from continue/next/step/stepout
even if we don't have a source file for the instruction that we are
stopped on.

This is mostly important on macOS where a SIGSEGV will cause 'continue'
to fail with a 'bad access' error (see #852) and the output can be
confusing.

Fixes #1244