pkg/config: try to use XDG_CONFIG_HOME when present
* config/config.go

This change checks for the environmental variable XDG_CONFIG_HOME
before storing the configuration file in the home directory.

Closes #1454

* config/config.go

This change attempts to move the previous config file location to the
new XDG_CONFIG_HOME compliant location on linux systems or when
XDG_CONFIG_HOME is set.

This change removes the old config file location if the file is
successfully moved.

* config/config.go

Consolidate messages about config moving and clean up into one.
Remove extraneous newlines.

* pkg/config: config path error messages print to stderr

This change updates the output location of error messages related to
the config path. This will prevent clients from getting unexpected output
when an error occurs.

* pkg/config: Update path used when returning unable to move config error

This change fixes the error message when Delve was unable to move the config. Previously returned a FileInfo object not the file path.

* pkg/config: Remove check for stderr when printing error messages

* pkg/config: Move success message for moving config to stderr

This prevents the success message from being printed to stdout and
breaking frontends.

* pkg/config: Rename variable to be more descriptive